refactor: Refactor adjustement
